+++ src/main/kotlin/util/Constants.kt
@@ -1,0 +1,12 @@
+package util
+
+object Constants {
+    const val DELIMITER_CAR_INPUT = ","
+    const val INDICATOR_CAR_POSITION = "-"
+    const val DELIMITER_WINNER_OUTPUT = ", "
+    const val CONDITION_CAR_MAX_LENGTH = 5
+    const val CONDITION_CAR_MIN_COUNT = 1
+    const val CONDITION_RANDOM_MIN = 0
+    const val CONDITION_RANDOM_MAX = 9
+    const val CONDITION_MOVE_THRESHOLD = 4
+}
