+++ src/main/kotlin/util/Messages.kt
@@ -1,0 +1,8 @@
+package util
+
+object Messages {
+    const val MESSAGE_INPUT_CAR_NAMES = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분)."
+    const val MESSAGE_INPUT_RACE_COUNT = "시도할 횟수는 몇 회인가요?"
+    const val MESSAGE_OUTPUT_RACE_RESULT = "실행 결과"
+    const val MESSAGE_OUTPUT_WINNER = "최종 우승자: "
+}
